Fix lint error part 6
