+++ cmd/soroban-rpc/internal/methods/get_events.go
@@ -7,15 +7,15 @@ import (
 	"strings"
 	"time"
 
-	"github.com/stellar/go/support/log"
-	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/db"
-	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/ledgerbucketwindow"
-
 	"github.com/creachadair/jrpc2"
 
 	"github.com/stellar/go/strkey"
 	"github.com/stellar/go/support/errors"
+	"github.com/stellar/go/support/log"
 	"github.com/stellar/go/xdr"
+
+	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/db"
+	"github.com/stellar/soroban-rpc/cmd/soroban-rpc/internal/ledgerbucketwindow"
 )
 
 type eventTypeSet map[string]interface{}
